分离chatwidget数据和界面

+++ ui/LogInWidget/LogInWidget.cpp
@@ -71,7 +71,7 @@ void LogInWidget::onSignalLoginResultRecv(const QString& msg)
     if (loginReplyData.m_bLoginInResult)
     {
         //打开对应的聊天界面
-        ChatWidget* ptrChatWidget = new ChatWidget(m_strUserId.toInt());
+        ChatWidget* ptrChatWidget = new ChatWidget(m_strUserId);
         //该界面隐藏,一会后析构
         hide();
         close();
